Move verse-related command groups into their own folder.
